// server/src/types.ts
export type Day = 'M' | 'Tu' | 'W' | 'Th' | 'F' | 'Sa' | 'Su' | 'TBA';
// ...
export interface MeetingInput {
    day: Day;
    startTime: string;
    endTime: string;
    location: string;
    sectionId: number;
}